add user controller-registration

+++ src/utils/cloudinary.js
@@ -1,22 +1,34 @@
+import "../config.js";
 import {v2 as cloudinary} from 'cloudinary'
+import fs from "fs"
+
+// console.log("Cloud Name:", process.env.CLOUDINARY_CLOUD_NAME);
+// console.log("API Key:", process.env.CLOUDINARY_API_KEY);
+// console.log("API Secret:", process.env.CLOUDINARY_API_SECRET);
 
 cloudinary.config({
-    cloud_name:process.env.CLOUDINARY_API_KEY,
-    api_key:process.env.CLOUDINARY_API_SECRET,
-    cloud_name:process.env.CLOUDINARY_CLOUD_NAME,
+  cloud_name:process.env.CLOUDINARY_CLOUD_NAME,
+  api_key:process.env.CLOUDINARY_API_KEY,
+  api_secret:process.env.CLOUDINARY_API_SECRET,
 })
 
 const UploadOnCloudinary = async (imagePath) => {
-    try {
+  try {
       if(!imagePath)return null
       const result = await cloudinary.uploader.upload(imagePath,{
         resource_type:"auto"
       });
-      console.log("file has been uploaded",result.url);
+      // console.log("file has been uploaded",result);
+      fs.unlinkSync(imagePath);
       return result;
     } catch (error) {
+      console.log("Cloudinary Error:", error);
+
+      if (imagePath && fs.existsSync(imagePath)) {
         fs.unlinkSync(imagePath);
-      console.error(error);
+      }
+
+      return null;
     }
 };
